requirements.txt
Flask==2.3.3
pyserial==3.5
requests==2.31.0
pyyaml==6.0
